implement min. Changed code to work with doubles instead of ints

+++ includes/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: pcunha <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/15 20:42:01 by pcunha            #+#    #+#             */
-/*   Updated: 2021/07/17 20:24:18 by pcunha           ###   ########.fr       */
+/*   Updated: 2021/07/20 18:23:20 by pcunha           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,14 +35,14 @@ int		is_integer(char *s);
 double	*populate_array(int argc, char **array_bi);
 void	q_sort(double *array, int left, int right);
 void	check_repeated(double *array, int n_array);
-t_dlist	*ft_dlst_push(int *value, t_dlist *list);
+t_dlist	*ft_dlst_push(double *value, t_dlist *list);
 void	u_print_dlst(t_dlist *list);
 void	u_free_dlst(t_dlist *list);
 int		len_dlst(t_dlist *list);
 t_dlist	*last_dlst(t_dlist *list);
 t_dlist	*first_dlst(t_dlist *list);
 void	u_print_stack(t_dlist *list, char c);
-t_dlist	*push_array_into_stack(int *array, int n_array, t_dlist *list);
+t_dlist	*push_array_into_stack(double *array, int n_array, t_dlist *list);
 void	push(t_dlist **to, t_dlist **from);
 void	rotate(t_dlist **a);
 void	reverse_rotate(t_dlist **a);
